pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "traceroot"
version = "0.0.3a6"
description = "A clean, principled wrapper around OpenTelemetry, AWS CloudWatch, and AWS X-Ray for enhanced debugging experience"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "TraceRoot Team"},
]
keywords = ["artificial-intelligence", "agent-systems", "opentelemetry", "tracing", "logging", "debugging"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Logging",
    "Topic :: System :: Monitoring",
]
dependencies = [
    "opentelemetry-api==1.34.1",
    "opentelemetry-sdk==1.34.1",
    "opentelemetry-exporter-otlp==1.34.1",
    "opentelemetry-exporter-otlp-proto-common==1.34.1",
    "opentelemetry-exporter-otlp-proto-grpc==1.34.1",
    "opentelemetry-exporter-otlp-proto-http==1.34.1",
    "opentelemetry-instrumentation==0.55b1",
    "opentelemetry-instrumentation-asgi==0.55b1",
    "opentelemetry-instrumentation-fastapi==0.55b1",
    "opentelemetry-proto==1.34.1",
    "opentelemetry-sdk-extension-aws==2.1.0",
    "opentelemetry-propagator-aws-xray==1.0.2",
    "opentelemetry-semantic-conventions==0.55b1",
    "opentelemetry-util-http==0.55b1",
    "watchtower==3.4.0",
    "pandas>=1.5.0",
    "PyYAML==6.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black",
    "flake8",
    "mypy",
]
fastapi = [
    "fastapi==0.115.12",
    "uvicorn==0.34.3",
    "pre-commit==4.2.0",
    "pytest==8.4.0",
    "httpx==0.27.0",
    "numpy",
    "opentelemetry-instrumentation-fastapi==0.55b1",
]

[project.urls]
Homepage = "https://traceroot.ai/"
Documentation = "https://docs.traceroot.ai"
Repository = "https://github.com/traceroot-ai/traceroot-sdk"
"Bug Reports" = "https://github.com/traceroot-ai/traceroot-sdk/issues"

[tool.setuptools]
include-package-data = true
zip-safe = false

[tool.setuptools.packages.find]
where = ["."]
include = ["traceroot*", "integrations*"]
exclude = ["tests*"] 